Extract ArrowRecordExtractor type conversion into shared ArrowToPinotTypeConverter
Pure refactor — no behavior change. Move the schema-driven Arrow → Pinot value
conversion (originally introduced in #18434) from ArrowRecordExtractor into a
new public static utility ArrowToPinotTypeConverter so it can be reused by
column-major ColumnReader implementations that wrap Arrow vectors.

ArrowToPinotTypeConverter exposes a single entry point:

    public static Object toPinotValue(Field field, Object value, boolean extractRawTimeValues)

with all per-type helpers (convertTimestamp / Date / Time / List / Map /
Struct / byRuntimeType) reduced to private static methods. The
extractRawTimeValues flag — previously an ArrowRecordExtractor instance
field — is threaded through as a method parameter so the converter remains
stateless.

ArrowRecordExtractor.convert(Field, Object) is removed; the per-row dispatch
in extract() now calls ArrowToPinotTypeConverter.toPinotValue() directly.
All existing tests in pinot-arrow continue to pass unchanged:

- ArrowRecordReaderTest (2 / 2)
- ArrowRecordExtractorTest (54 / 54)
- ArrowMessageDecoderTest (7 / 7)

This precursor lets a follow-up ColumnReader implementation reuse the
conversion without duplicating ~200 lines of schema-driven dispatch.
